### 2025-10-29 #9320 @editorialbot
[PRE REVIEW]: Jaxion: A JAX package for Fuzzy Dark Matter
**Submitting author:** @pmocz (<a href="http://orcid.org/0000-0001-6631-2566">Philip Mocz</a>)
**Repository:** https://github.com/JaxionProject/jaxion
**Branch with paper.md** (empty if default branch):
**Version:** 0.0.3
