### 2025-10-29 dfbd694
fix distributed vis, improve sbatch files
