+++ examples/tidal_stripping/sbatch_rusty_distributed.sh
@@ -6,7 +6,7 @@
 #SBATCH --constraint=h100
 #SBATCH --nodes=2
 #SBATCH --ntasks-per-node=4
-#SBATCH --gpus-per-node=4
+#SBATCH --gpus-per-task=1
 #SBATCH --cpus-per-task=8
 #SBATCH --mem=80G
 #SBATCH --time=00-00:20
