### 2025-10-29 fe66269
rename logo to logo_inverse_problem for clarity
