+++ scripts/viswall.sh
@@ -14,7 +14,7 @@ convert ../examples/dynamical_friction/movie.gif  -coalesce input2_frame_%04d.pn
 convert ../examples/heating_gas/movie.gif  -coalesce input3_frame_%04d.png
 convert ../examples/heating_stars/movie.gif  -coalesce input4_frame_%04d.png
 convert ../examples/kinetic_condensation/movie.gif -coalesce input5_frame_%04d.png
-convert ../examples/logo/movie.gif  -coalesce input6_frame_%04d.png
+convert ../examples/logo_inverse_problem/movie.gif  -coalesce input6_frame_%04d.png
 convert ../examples/soliton_binary_merger/movie.gif  -coalesce input7_frame_%04d.png
 convert ../examples/soliton_merger/movie.gif  -coalesce input8_frame_%04d.png
 convert ../examples/tidal_stripping/movie.gif  -coalesce input9_frame_%04d.png
