+++ examples/logo_inverse_problem/README.md
@@ -1,4 +1,4 @@
-# Logo
+# Logo Inverse Problem
 
 Inverse problem: Find initial conditions for velocity (dm+stars) to achieve target density at t=1
 
@@ -7,7 +7,7 @@ Philip Mocz (2025)
 Usage:
 
 ```console
-python logo.py
+python logo_inverse_problem.py
 ```
 
 Takes around 200 seconds to run on my macbook (cpu).
