### 2025-10-29 6aae138
coverage for analysis functions
