+++ tests/test_analysis.py
@@ -21,4 +21,4 @@ def test_quantum_velocity_and_radial_power_spectrum():
 
     Pf, _, _ = radial_power_spectrum(vx, kx, ky, kz, box_size)
 
-    assert jnp.max(Pf) == pytest.approx(8244.606, abs=1e-4)
+    assert jnp.max(Pf) == pytest.approx(8244.606, rel=1e-4)
