### 2025-10-30 b3fb3a5
remove redundant code in logo problem, update gif
