### 2025-11-04 99d6b4d
fix  pyproject.toml include
