### 2025-11-05 #9320 @arfon
:wave: @pmocz – could you take a look a [this list of potential reviewers](https://reviewers.joss.theoj.org/lookup) and identify a few people who would be good candidates to review this submission?
