+++ install.sh
@@ -1,2 +1,4 @@
+#!/bin/bash
+
 pip uninstall jaxion -y
 pip install -e .
