### 2025-11-05 b6cf391
adding frontier scripts
