### 2025-11-05 4dbd836
update frontier distributed run settings
