### 2025-11-05 c9dab0d
add plot script
