+++ examples/soliton_binary_merger/plot_timing.py
@@ -47,7 +47,7 @@ mcups = np.array(
 num_nodes = np.array([1, 1, 1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 2, 1, 1, 1, 2])
 num_gpus = np.array([1, 1, 1, 1, 2, 4, 8, 16, 1, 2, 4, 8, 16, 2, 4, 8, 16])
 
-yy = np.array([0.01, 0.001, 0.1, 1, 10])
+yy = np.array([0.001, 0.01, 0.1, 1, 10])
 
 
 def main():
