### 2025-11-06 8bea12c
adding frontier debug example
