### 2025-11-06 ed5c2a0
update debug tests
