+++ examples/soliton_binary_merger/sbatch_frontier.sh
@@ -14,7 +14,7 @@ module purge
 module load PrgEnv-gnu/8.6.0
 module load rocm/6.2.4
 module load craype-accel-amd-gfx90a
-module load miniforge3/23.11.0-0ls
+module load miniforge3/23.11.0-0
 
 export PYTHONUNBUFFERED=TRUE
 
