### 2025-11-07 285b186
fix install script
