### 2025-11-07 f6d5151
added small test
