### 2025-11-07 e3fd701
decrease test times
