+++ examples/debug/README.md
@@ -23,19 +23,19 @@ cd jaxion
 
 ## Submit jobs
 
-Examples to use for testing/debugging are in the `examples/debug` folder:
+Examples to use for testing/debugging are in the `examples/debug/` folder:
 
 ```console
 cd examples/debug
 ```
 
-Submit a Large (L) job (resolution 2046^3) on 16 nodes / 64 GPUs:
+Submit a Large (L) job (resolution 2048^3) on 16 nodes / 64 GPUs:
 
 ```console
 sbatch sbatch_frontier_L.sh
 ```
 
-(this test should run in about 23 minutes.)
+(this test should run in about 160 seconds.)
 
 Submit an Extra Large (XL) job (resolution 4096^3) on 64 nodes / 512 GPUs:
 
@@ -43,19 +43,27 @@ Submit an Extra Large (XL) job (resolution 4096^3) on 64 nodes / 512 GPUs:
 sbatch sbatch_frontier_XL.sh
 ```
 
-This test currently fails with:
+This test currently fails immediately with:
 
 ```console
 E1105 14:17:03.123606  763404 pjrt_stream_executor_client.cc:3045] Execution of replica 0 failed: INVALID_ARGUMENT: CliqueIds size must be 1 for NCCL communicator initialization
 ```
 
-Also, as a refernce, we include a Small (S) job (resolution 512^3) on 1 nodes / 1 GPU:
+Also, as a reference, we include a Small (S) job (resolution 512^3) on 1 nodes / 1 GPU:
 
 ```console
 sbatch sbatch_frontier_S.sh
 ```
 
-which should take under 30 seconds to complete.
+which takes around 4.5 seconds to complete.
+
+And a Medium (M) job resolution (1024^3) on 2 nodes / 8 GPUs:
+
+```console
+sbatch sbatch_frontier_M.sh
+```
+
+which takes around 78 seconds to complete.
 
 ## Notes
 
