### 2025-11-07 b667995
adding references
