+++ examples/heating_stars/README.md
@@ -12,8 +12,14 @@ python heating_stars.py --res <resolution_multiplier>
 
 Takes around 11 (res=1) seconds to run on my macbook (cpu).
 
+
 ## Simulation snapshots
 
 <div style="display:flex;flex-wrap:wrap;gap:8px">
   <img src="dm100.png" alt="dm100" width="45%"/>
 </div>
+
+
+## References
+
+[Church, B.; Mocz, P.; Ostriker, J.P.; Heating of Milky Way disc Stars by Dark Matter Fluctuations in Cold Dark Matter and Fuzzy Dark Matter Paradigms. MNRAS (2019)](https://ui.adsabs.harvard.edu/abs/2019MNRAS.485.2861C)
