### 2025-11-07 3e6f681
adding aws-ofi-rccl to frontier setup
