+++ examples/debug/plot_timing.py
@@ -12,6 +12,7 @@ resolutions = np.array(
         #        8192,
     ]
 )
+# run time for 10 steps
 runtime = np.array(
     [
         4.5,
@@ -33,13 +34,13 @@ runtime_no_awsofirccl = np.array(
 num_nodes = np.array([1, 2, 16])  # , 128, 1024])
 num_gpus = np.array([1, 8, 64])  # , 512, 4096])
 
-yy = np.array([0.01, 0.1, 1])
+yy = np.array([0.1, 1, 10])
 
 
 def main():
     # Compute billion cell updates per second
-    bcups = (resolutions**3) / (runtime * 1.0e9)
-    bcups_no_awsofirccl = (resolutions**3) / (runtime_no_awsofirccl * 1.0e9)
+    bcups = 10 * (resolutions**3) / (runtime * 1.0e9)
+    bcups_no_awsofirccl = 10 * (resolutions**3) / (runtime_no_awsofirccl * 1.0e9)
 
     plt.figure(figsize=(4, 4))
     plt.plot(
@@ -57,9 +58,9 @@ def main():
     for i, (res, bcup, nodes, gpus) in enumerate(
         zip(resolutions, bcups, num_nodes, num_gpus)
     ):
-        plt.text(res * 0.9, bcup / 1.3, f"# nodes={nodes}", fontsize=7, va="center")
+        plt.text(res * 0.85, bcup / 1.22, f"# nodes={nodes}", fontsize=7, va="center")
         plt.text(
-            res * 0.9, bcup / (1.3) ** 2, f"# gpus={gpus}", fontsize=7, va="center"
+            res * 0.85, bcup / (1.22) ** 2, f"# gpus={gpus}", fontsize=7, va="center"
         )
     plt.xscale("log")
     plt.yscale("log")
@@ -70,7 +71,7 @@ def main():
     plt.gca().set_xticks([], minor=True)
     plt.yticks(yy, labels=[f"{y}" for y in yy])
     plt.xlim(resolutions[0] / 1.3, resolutions[-1] * 1.3)
-    plt.legend(fontsize=8)
+    plt.legend(fontsize=8, loc="upper left")
     plt.tight_layout()
     plt.savefig("timing.png", dpi=300)
     plt.show()
