+++ docs/pages/citing.rst
@@ -3,14 +3,7 @@ Citing
 
 If you use this software, please cite it as below.
 
-.. code-block:: bibtex
-
-  @software{Mocz_Jaxion_2025,
-     author = {Mocz, Philip},
-        doi = {10.5281/zenodo.17438467},
-      month = oct,
-      title = {{Jaxion}},
-        url = {https://github.com/JaxionProject/jaxion},
-    version = {0.0.4},
-       year = {2025}
-  }
+.. literalinclude:: ../../README.md
+  :start-at: Mocz_Jaxion_2025
+  :end-before: ```
+  :language: latex
