+++ jaxion/gravity.py
@@ -5,6 +5,6 @@ import jaxdecomp as jd
 
 
 def calculate_gravitational_potential(rho, k_sq, G, rho_bar):
-    Vhat = -jd.fft.pfft3d(4.0 * jnp.pi * G * (rho - rho_bar)) / (k_sq + (k_sq == 0))
-    V = jnp.real(jd.fft.pifft3d(Vhat))
+    V_hat = -jd.fft.pfft3d(4.0 * jnp.pi * G * (rho - rho_bar)) / (k_sq + (k_sq == 0))
+    V = jnp.real(jd.fft.pifft3d(V_hat))
     return V
