+++ jaxion/simulation.py
@@ -282,6 +282,7 @@ class Simulation:
         dx = self.dx
         m_per_hbar = self.m_per_hbar
         box_size = self.box_size
+        num_cells = self.resolution**3
 
         dt_fac = 1.0
         dt_kin = dt_fac * (m_per_hbar / 6.0) * (dx * dx)
@@ -290,8 +291,16 @@ class Simulation:
         t_span = t_end - t_start
         state["t"] = t_start
 
+        use_quantum = self.params["physics"]["quantum"]
+        use_gravity = self.params["physics"]["gravity"]
+        use_hydro = self.params["physics"]["hydro"]
+        use_particles = self.params["physics"]["particles"]
+        use_cosmology = self.params["physics"]["cosmology"]
+        use_external_potential = self.params["physics"]["external_potential"]
+        save = self.params["output"]["save"]
+
         # cosmology
-        if self.params["physics"]["cosmology"]:
+        if use_cosmology:
             z_start = self.params["time"]["start"]
             z_end = self.params["time"]["end"]
             omega_matter = self.params["cosmology"]["omega_matter"]
@@ -317,12 +326,12 @@ class Simulation:
         k_sq = None
 
         # Fourier space variables
-        if self.params["physics"]["gravity"] or self.params["physics"]["quantum"]:
+        if use_gravity or use_quantum:
             kx, ky, kz = self.kgrid
             k_sq = kx**2 + ky**2 + kz**2
 
         # Checkpointer
-        if self.params["output"]["save"]:
+        if save:
             options = ocp.CheckpointManagerOptions()
             checkpoint_dir = os.path.join(os.getcwd(), self.params["output"]["path"])
             path = os.path.join(os.getcwd(), checkpoint_dir)
@@ -334,43 +343,36 @@ class Simulation:
 
         def _kick(state, kx, ky, kz, k_sq, dt):
             # Kick (half-step)
-            if (
-                self.params["physics"]["gravity"]
-                and self.params["physics"]["external_potential"]
-            ):
+            if use_gravity and use_external_potential:
                 V = self._calc_grav_potential(state, k_sq) + state["V_ext"]
-            elif self.params["physics"]["gravity"]:
+            elif use_gravity:
                 V = self._calc_grav_potential(state, k_sq)
-            elif self.params["physics"]["external_potential"]:
+            elif use_external_potential:
                 V = state["V_ext"]
 
-            if (
-                self.params["physics"]["gravity"]
-                or self.params["physics"]["external_potential"]
-            ):
-                if self.params["physics"]["quantum"]:
+            if use_gravity or use_external_potential:
+                if use_quantum:
                     state["psi"] = quantum_kick(state["psi"], V, m_per_hbar, dt)
-                if self.params["physics"]["hydro"]:
+                if use_hydro:
                     state["vx"], state["vy"], state["vz"] = hydro_accelerate(
                         state["vx"], state["vy"], state["vz"], V, kx, ky, kz, dt
                     )
-                if self.params["physics"]["particles"]:
+                if use_particles:
                     state["vel"] = particles_accelerate(
                         state["vel"], state["pos"], V, kx, ky, kz, dx, dt
                     )
 
         def _drift(state, k_sq, dt):
             # Drift (full-step)
-            if self.params["physics"]["quantum"]:
+            if use_quantum:
                 state["psi"] = quantum_drift(state["psi"], k_sq, m_per_hbar, dt)
-            if self.params["physics"]["hydro"]:
+            if use_hydro:
                 state["rho"], state["vx"], state["vy"], state["vz"] = hydro_fluxes(
                     state["rho"], state["vx"], state["vy"], state["vz"], dt, dx, c_sound
                 )
[…]
