### 2025-11-12 9bdc1ee
add self-interaction
