+++ docs/pages/parameters.rst
@@ -3,5 +3,5 @@ Parameters
 
 Jaxion allows the user to configure the simulation via the following set of parameters:
 
-.. literalinclude:: ../../jaxion/params_default.json
+.. literalinclude:: ../../jaxion/defaults.json
   :language: json
