+++ examples/two_field/two_field.py
@@ -1,0 +1,200 @@
+import jax
+import jax.numpy as jnp
+import numpy as np
+import jaxdecomp as jd
+import argparse
+import sys
+
+sys.path.append("../../")  # XXX
+import jaxion
+
+# switch on for double precision
+# jax.config.update("jax_enable_x64", True)
+
+"""
+Merge solitons to form an idealized Two-Field Fuzzy Dark Matter halo
+Demonstrates adding custom fields
+
+Philip Mocz (2025)
+
+Usage:
+  python two_field.py --res <resolution_multiplier>
+"""
+
+
+def set_up_simulation(resolution_multiplier):
+    # Parameters added/changed from default values
+    params = {
+        "physics": {
+            "quantum": False,  # use custom fields instead for psi1, psi2
+            "gravity": True,
+        },
+        "domain": {
+            "resolution_multiplier": resolution_multiplier,
+        },
+        "output": {
+            "path": f"./checkpoints{resolution_multiplier}/",
+        },
+    }
+
+    # Initialize the simulation
+    sim = jaxion.Simulation(params)
+
+    # Set initial conditions (randomly placed solitons)
+    m_22_1 = 1.0
+    m_22_2 = 2.0
+    hbar = jaxion.constants["reduced_planck_constant"]
+    eV = jaxion.constants["electron_volt"]
+    c = jaxion.constants["speed_of_light"]
+    m1 = m_22_1 * 1.0e-22 * eV / c**2
+    m2 = m_22_2 * 1.0e-22 * eV / c**2
+    m1_per_hbar = m1 / hbar
+    m2_per_hbar = m2 / hbar
+
+    box_size = sim.params["domain"]["box_size"]
+    nx = sim.resolution
+    xx, yy, zz = sim.grid
+
+    def rho_soliton(r, r_soliton, m_22):
+        return (
+            1.9e7 * m_22**-2 * r_soliton**-4 / (1.0 + 0.091 * (r / r_soliton) ** 2) ** 8
+        )
+
+    np.random.seed(17)
+    n_solitons = 8
+
+    rho1 = np.zeros((nx, nx, nx), dtype=complex)
+    rho2 = np.zeros((nx, nx, nx), dtype=complex)
+    for _ in range(n_solitons):
+        r_soliton = (0.05 + 0.03 * np.random.rand()) * box_size
+        x_soliton = (0.25 + 0.5 * np.random.rand()) * box_size
+        y_soliton = (0.25 + 0.5 * np.random.rand()) * box_size
+        z_soliton = (0.25 + 0.5 * np.random.rand()) * box_size
+
+        r = jnp.sqrt(
+            (xx - x_soliton) ** 2 + (yy - y_soliton) ** 2 + (zz - z_soliton) ** 2
+        )
+        rho1 += rho_soliton(r, r_soliton, m_22_1)
+        rho2 += rho_soliton(r, r_soliton, m_22_2)
+
+    sim.state["psi1"] = jnp.array(jnp.sqrt(rho1))
+    sim.state["psi2"] = jnp.array(jnp.sqrt(rho2))
+
+    # Define custom functions
+    def custom_density(state):
+        return jnp.abs(state["psi1"]) ** 2 + jnp.abs(state["psi2"]) ** 2
+
+    def custom_kick(state, V, dt):
+        state["psi1"] = jnp.exp(-1j * m1_per_hbar * dt * V) * state["psi1"]
+        state["psi2"] = jnp.exp(-1j * m2_per_hbar * dt * V) * state["psi2"]
+
+        return state
+
+    def custom_drift(state, k_sq, dt):
[…]
