+++ examples/two_field/README.md
@@ -1,0 +1,27 @@
+# Two-Field
+
+Demo for adding custom fields: Two-field fuzzy dark matter
+Merge solitons in two-field FDM to form a halo
+
+Philip Mocz (2025)
+
+Usage:
+
+```console
+python two_field.py --res <resolution_multiplier>
+```
+
+Takes around 6 (res=1) seconds to run on my macbook (cpu).
+
+
+## Simulation snapshots
+
+<div style="display:flex;flex-wrap:wrap;gap:8px">
+  <img src="rho1_070.png" alt="rho1_070" width="45%"/>
+  <img src="rho2_070.png" alt="rho2_070" width="45%"/>
+</div>
+
+
+## References
+
+[Luu, H.N.; Mocz, P.; Vogelsberger, M.; Nested solitons in two-field fuzzy dark matter. MNRAS (2024)](https://ui.adsabs.harvard.edu/abs/2024MNRAS.527.4162L)
