### 2025-11-17 3469f47
bugfix in rho_bar particle binning, and add accretion
