+++ jaxion/visualization.py
@@ -82,6 +82,14 @@ def plot_sim(state, checkpoint_dir, i, params):
                 vmax=vmax,
                 extent=[0, nx, 0, nx],
             )
+            if params["physics"]["particles"]:
+                # draw particles
+                box_size = params["domain"]["box_size"]
+                sx = (state["pos"][:, 0] / box_size) * nx
+                sy = (state["pos"][:, 1] / box_size) * nx
+                plt.plot(
+                    sx, sy, color="red", marker=".", linestyle="None", markersize=5
+                )
             ax.set_aspect("equal")
             ax.get_xaxis().set_visible(False)
             ax.get_yaxis().set_visible(False)
