+++ examples/black_hole_accretion/README.md
@@ -1,0 +1,26 @@
+# Black Hole Accretion
+
+Accrete gas onto a black hole particle
+
+Philip Mocz (2025)
+
+Usage:
+
+```console
+python black_hole_accretion.py --res <resolution_multiplier>
+```
+
+Takes around 20 (res=1) seconds to run on my macbook (cpu).
+
+
+## Simulation snapshots
+
+<div style="display:flex;flex-wrap:wrap;gap:8px">
+  <img src="dm100.png" alt="dm100" width="45%"/>
+  <img src="gas100.png" alt="gas100" width="45%"/>
+</div>
+
+
+## References
+
+

[…]
