+++ examples/black_hole_accretion/black_hole_accretion.py
@@ -1,7 +1,4 @@
-import jax
 import jax.numpy as jnp
-import numpy as np
-import jaxdecomp as jd
 import argparse
 import sys
 
