### 2025-11-18 b711d89
add time safety fac as user parameter
