### 2025-11-18 bf6e82d
add bh mass plot, load from snap number
