+++ examples/black_hole_accretion/sbatch_rusty.sh
@@ -1,0 +1,21 @@
+#!/usr/bin/bash
+#SBATCH --job-name=black_hole_accretion
+#SBATCH --output=slurm-%j.out
+#SBATCH --error=slurm-%j.err
+#SBATCH --partition gpu
+#SBATCH --constraint=h100
+#SBATCH --nodes=1
+#SBATCH --ntasks-per-node=1
+#SBATCH --gpus-per-task=1
+#SBATCH --cpus-per-task=8
+#SBATCH --mem=80G
+#SBATCH --time=00-02:00
+
+module purge
+module load python/3.11
+
+export PYTHONUNBUFFERED=TRUE
+
+source $VENVDIR/jaxion-venv/bin/activate
+
+srun python black_hole_accretion.py --res=4
