### 2025-11-22 9a0ef5d
add callback, move plot_sim after async save
