### 2025-11-25 ba816f5
upgrade to jax 0.6.0, fix frontier setup
