+++ examples/debug/sbatch_frontier_L.sh
@@ -4,15 +4,12 @@
 #SBATCH --output=slurm-%j.out
 #SBATCH --error=slurm-%j.err
 #SBATCH --partition batch
-#SBATCH --nodes=16
-#SBATCH --ntasks-per-node=4
-#SBATCH --gpus-per-task=1
-#SBATCH --cpus-per-task=8
-#SBATCH --time=00-00:30
+#SBATCH --nodes=8
+#SBATCH --time=00-00:05
 
 module purge
 module load PrgEnv-gnu/8.6.0
-module load rocm/6.2.4
+module load rocm/7.0.2
 module load craype-accel-amd-gfx90a
 module load miniforge3/23.11.0-0
 
@@ -21,6 +18,9 @@ export LD_LIBRARY_PATH=/lustre/orion/ast231/scratch/pmocz/aws-ofi-rccl/lib:$LD_L
 export FI_MR_CACHE_MONITOR=kdreg2
 export PYTHONUNBUFFERED=TRUE
 
-conda activate jaxion-venv
+# see https://docs.olcf.ornl.gov/software/analytics/pytorch_frontier.html
+# for more optimization tips
 
-srun --cpu-bind=cores python debug.py --res=64 --distributed
+conda activate /lustre/orion/scratch/pmocz/ast231/venvs/jaxion-venv
+
+srun -N$SLURM_NNODES --ntasks-per-node=8 --cpus-per-task=7 --gpu-bind=closest --cpu-bind=cores python debug.py --res=64 --distributed
