+++ examples/debug/sbatch_frontier_S.sh
@@ -5,14 +5,11 @@
 #SBATCH --error=slurm-%j.err
 #SBATCH --partition batch
 #SBATCH --nodes=1
-#SBATCH --ntasks-per-node=1
-#SBATCH --gpus-per-task=1
-#SBATCH --cpus-per-task=8
 #SBATCH --time=00-00:05
 
 module purge
 module load PrgEnv-gnu/8.6.0
-module load rocm/6.2.4
+module load rocm/7.0.2
 module load craype-accel-amd-gfx90a
 module load miniforge3/23.11.0-0
 
@@ -21,6 +18,6 @@ export LD_LIBRARY_PATH=/lustre/orion/ast231/scratch/pmocz/aws-ofi-rccl/lib:$LD_L
 export FI_MR_CACHE_MONITOR=kdreg2
 export PYTHONUNBUFFERED=TRUE
 
-conda activate jaxion-venv
+conda activate /lustre/orion/scratch/pmocz/ast231/venvs/jaxion-venv
 
-srun --cpu-bind=cores python debug.py --res=16
+srun -N$SLURM_NNODES --ntasks-per-node=1 --cpus-per-task=7 --gpu-bind=closest --cpu-bind=cores python debug.py --res=16
