+++ examples/debug/plot_timing.py
@@ -8,31 +8,23 @@ resolutions = np.array(
         512,
         1024,
         2048,
-        #        4096,
+        4096,
         #        8192,
     ]
 )
 # run time for 10 steps
 runtime = np.array(
     [
-        4.5,
-        12,
-        27,
-        #        999,
-        #        999,
-    ]
-)
-runtime_no_awsofirccl = np.array(
-    [
-        4.5,
-        78,
-        160,
-        #        999,
+        3.8,
+        6.3,
+        136.1,
+        174.3,
         #        999,
     ]
 )
-num_nodes = np.array([1, 2, 16])  # , 128, 1024])
-num_gpus = np.array([1, 8, 64])  # , 512, 4096])
+
+num_nodes = np.array([1, 1, 8, 64])  # 512])
+num_gpus = np.array([1, 8, 64, 512])  # 4096])  # really # GCDs (=2 * # gpus)
 
 yy = np.array([0.1, 1, 10])
 
@@ -40,7 +32,7 @@ yy = np.array([0.1, 1, 10])
 def main():
     # Compute billion cell updates per second
     bcups = 10 * (resolutions**3) / (runtime * 1.0e9)
-    bcups_no_awsofirccl = 10 * (resolutions**3) / (runtime_no_awsofirccl * 1.0e9)
+    #bcups_no_awsofirccl = 10 * (resolutions**3) / (runtime_no_awsofirccl * 1.0e9)
 
     plt.figure(figsize=(4, 4))
     plt.plot(
@@ -48,12 +40,6 @@ def main():
         bcups,
         marker="s",
         label="MI250X",
-    )
-    plt.plot(
-        resolutions,
-        bcups_no_awsofirccl,
-        marker="o",
-        label="MI250X (no aws-ofi-rccl)",
     )
     for i, (res, bcup, nodes, gpus) in enumerate(
         zip(resolutions, bcups, num_nodes, num_gpus)
