### 2025-11-25 641a86c
ruff
