### 2025-11-26 f40aa83
rusty switch to python 3.12.9, update sbatch files with gpu-bind=none
