+++ scripts/setup_rusty_venv.sh
@@ -10,7 +10,7 @@ fi
 rm -fr $VENVDIR/jaxion-venv
 
 module purge
-module load python/3.11
+module load python/3.12.9
 python -m venv --system-site-packages $VENVDIR/jaxion-venv
 source $VENVDIR/jaxion-venv/bin/activate
 pip install --upgrade pip
