+++ examples/debug/plot_timing.py
@@ -1,7 +1,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
 
-# Timings on Frontier & Rusty
+# Timings on Frontier & Rusty (H100 and H200)
 
 resolutions = np.array(
     [
@@ -23,7 +23,8 @@ runtime = np.array(
     ]
 )
 
-runtime_rusty = np.array([1.9, 3.0, np.nan, np.nan])  # , np.nan])
+runtime_rusty_h100 = np.array([2.2, 13.6, np.nan, np.nan])  # , np.nan])
+runtime_rusty_h200 = np.array([np.nan, np.nan, np.nan, np.nan])  # , np.nan])
 
 num_nodes = np.array([1, 1, 8, 64])  # 512])
 num_gpus = np.array([1, 8, 64, 512])  # 4096])  # really # GCDs (=2 * # gpus)
@@ -36,20 +37,27 @@ def main():
     nsteps = 10
     bcups = nsteps * (resolutions**3) / (runtime * 1.0e9)
     # bcups_no_awsofirccl = 10 * (resolutions**3) / (runtime_no_awsofirccl * 1.0e9)
-    bcups_rusty = nsteps * (resolutions**3) / (runtime_rusty * 1.0e9)
+    bcups_rusty_h100 = nsteps * (resolutions**3) / (runtime_rusty_h100 * 1.0e9)
+    bcups_rusty_h200 = nsteps * (resolutions**3) / (runtime_rusty_h200 * 1.0e9)
 
     plt.figure(figsize=(4, 4))
+    plt.plot(
+        resolutions,
+        bcups_rusty_h200,
+        marker="^",
+        label="H200 (Rusty)",
+    )
+    plt.plot(
+        resolutions,
+        bcups_rusty_h100,
+        marker="o",
+        label="H100 (Rusty)",
+    )
     plt.plot(
         resolutions,
         bcups,
         marker="s",
         label="MI250X (Frontier)",
-    )
-    plt.plot(
-        resolutions,
-        bcups_rusty,
-        marker="o",
-        label="H100 (Rusty)",
     )
     for i, (res, bcup, nodes, gpus) in enumerate(
         zip(resolutions, bcups, num_nodes, num_gpus)
