+++ docs/pages/tutorial1.rst
@@ -1,0 +1,201 @@
+Tutorial 1: Basic Simulation
+============================
+
+This tutorial will help you with the basics of setting up and running a simulation with Jaxion.
+It describes the example: `examples/soliton_gas_star <https://github.com/JaxionProject/jaxion/tree/main/examples/soliton_gas_star>`_
+
+Jaxion is designed for astrophysical scale simulations of fuzzy dark matter.
+As such, it uses units of:
+[L] = kpc, [V] = km/s, [M] = Msun
+(unless otherwise specified).
+All other units are derived from these base units;
+e.g., units of time are in [T] = [L]/[V] = kpc / (km/s) ~= 0.978 Gyr.
+
+First, in your Python script or Jupyter notebook, import the Jaxion library:
+
+.. code-block:: python
+
+    import jaxion
+
+It will also often be useful to import `jax.numpy <https://docs.jax.dev/en/latest/jax.numpy.html>`_:
+
+.. code-block:: python
+
+    import jax.numpy as jnp
+
+to set up initial conditions, which as JAX arrays.
+
+The basic steps are:
+
+(1) set simulation parameters,
+(2) create the simulation object,
+(3) set the initial conditions,
+(4) run the simulation.
+
+
+Setting Simulation Parameters
+-----------------------------
+
+The simulation parameters are set in a Python dictionary.
+If nothing is specified, the default values will be used.
+The default values can be found on the :doc:`Parameters <parameters>` page.
+
+Let's look at the parameters for this example:
+
+.. code-block:: python
+
+    params = {
+        "physics": {
+            "quantum": True,
+            "gravity": True,
+            "hydro": True,
+            "particles": True,
+        },
+        "domain": {
+            "box_size": 10.0,
+            "resolution_base": 32,
+            "resolution_multiplier": 2,
+        },
+        "time": {
+            "start": 0.0,
+            "end": 1.0,
+        },
+        "output": {
+            "path": f"./checkpoints",
+            "num_checkpoints": 100,
+            "save": True,
+        },
+        "quantum": {
+            "m_22": 1.0,
+        },
+        "hydro": {
+            "sound_speed": 40.0,
+        },
+        "particles": {
+            "num_particles": 1,
+            "particle_mass": 1.0e7
+        },
+    }
+
+In the ``"physics"`` section, we specify which physics modules to use.
+We enable the modules for the quantum wavefunction, self-gravity, hydrodynamics, and (star) particles.
+
+In the ``"domain"`` section, we set the size of the simulation box to 10 kpc,
+with a base resolution of 32 grid cells per side,
+and a resolution multiplier of 2 (meaning the effective resolution is 64 grid cells per side).
+
+In the ``"time"`` section, we set the simulation to start at time 0.0 and end at time 1.0 (in code units (kpc / (km/s))).
+
+In the ``"output"`` section, we specify the output path for saving simulation checkpoints,
+the number of checkpoints to save (100), and enable saving.
+
+In the ``"quantum"`` section, we set the fuzzy dark matter particle mass to m_22 = 1.0 (in units of 10^-22 eV).
+
+In the ``"hydro"`` section, we set the sound speed of the isothermal gas to 40.0 km/s.
+
+In the ``"particles"`` section, we specify that we want to include 1 star particle with a mass of 1.0e7 Msun.
+
+
+Creating the Simulation Object
+------------------------------
+
[…]
