### 2025-11-29 #9320 @jordydavelaar
Hi Arfon,I was traveling for work and the Thanksgiving holidays but are back now. I would be happy to take a lookBestJordyOn Nov 29, 2025, at 12:45 PM, Arfon Smith ***@***.***> wrote:﻿arfon left a comment (openjournals/joss-reviews#9320)
👋 @jmsexton03 @zingale @ttricco @TomGoffrey would any of you be willing to review this submission for JOSS? The submission under consideration is Jaxion: A JAX package for Fuzzy Dark Matter
Many thanks
Arfon
