### 2025-12-01 #9320 @cphyc
Hi,
Unfortunately, I cannot commit time at the moment for reviewing the paper, but I remain available for future ones.
