+++ examples/debug/sbatch_frontier_S.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/bash
 #SBATCH -A AST231
 #SBATCH --job-name=jaxionS
-#SBATCH --output=slurm-%j.out
-#SBATCH --error=slurm-%j.err
+#SBATCH --output=slurm-mi250xS-%j.out
+#SBATCH --error=slurm-mi250xS-%j.err
 #SBATCH --partition batch
 #SBATCH --nodes=1
 #SBATCH --time=00-00:05
