+++ examples/debug/sbatch_rusty_h200_M.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/bash
 #SBATCH --job-name=jaxionM
-#SBATCH --output=slurm-%j.out
-#SBATCH --error=slurm-%j.err
+#SBATCH --output=slurm-h200M-%j.out
+#SBATCH --error=slurm-h200M-%j.err
 #SBATCH --partition eval
 #SBATCH --constraint=h200
 #SBATCH --nodes=1
