+++ examples/debug/sbatch_frontier_M.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/bash
 #SBATCH -A AST231
 #SBATCH --job-name=jaxionM
-#SBATCH --output=slurm-%j.out
-#SBATCH --error=slurm-%j.err
+#SBATCH --output=slurm-mi250xM-%j.out
+#SBATCH --error=slurm-mi250xM-%j.err
 #SBATCH --partition batch
 #SBATCH --nodes=1
 #SBATCH --time=00-00:05
