### 2025-12-01 37e3389
sbatch partitions
