+++ examples/soliton_binary_merger/sbatch_frontier.sh
@@ -3,7 +3,7 @@
 #SBATCH --job-name=soliton_binary_merger
 #SBATCH --output=slurm-%j.out
 #SBATCH --error=slurm-%j.err
-#SBATCH --partition batch
+#SBATCH --partition=batch
 #SBATCH --nodes=8
 #SBATCH --time=00-00:20
 
