+++ examples/debug/sbatch_frontier_L.sh
@@ -3,7 +3,7 @@
 #SBATCH --job-name=jaxionL
 #SBATCH --output=slurm-mi250xL-%j.out
 #SBATCH --error=slurm-mi250xL-%j.err
-#SBATCH --partition batch
+#SBATCH --partition=batch
 #SBATCH --nodes=8
 #SBATCH --time=00-00:05
 
