+++ examples/debug/sbatch_rusty_h200_M.sh
@@ -2,7 +2,7 @@
 #SBATCH --job-name=jaxionM
 #SBATCH --output=slurm-h200M-%j.out
 #SBATCH --error=slurm-h200M-%j.err
-#SBATCH --partition eval
+#SBATCH --partition=eval
 #SBATCH --constraint=h200
 #SBATCH --nodes=1
 #SBATCH --ntasks-per-node=8
