+++ examples/debug/sbatch_frontier_XL.sh
@@ -3,7 +3,7 @@
 #SBATCH --job-name=jaxionXL
 #SBATCH --output=slurm-mi250xXL-%j.out
 #SBATCH --error=slurm-mi250xXL-%j.err
-#SBATCH --partition batch
+#SBATCH --partition=batch
 #SBATCH --nodes=64
 #SBATCH --time=00-00:10
 
