### 2025-12-04 523f95c
tutorials minor update
