### 2025-12-06 d0830a2
add notebook
