### 2025-12-09 ae9c46d
rm cpu-bind from frontier sbatch
