+++ examples/debug/sbatch_frontier_S.sh
@@ -20,4 +20,4 @@ export PYTHONUNBUFFERED=TRUE
 
 conda activate /lustre/orion/scratch/pmocz/ast231/venvs/jaxion-venv
 
-srun -N$SLURM_NNODES --ntasks-per-node=1 --cpus-per-task=7 --gpu-bind=closest --cpu-bind=cores python debug.py --res=16
+srun -N$SLURM_NNODES --ntasks-per-node=1 --cpus-per-task=7 --gpu-bind=closest python debug.py --res=16
