### 2025-12-09 f423e4b
update citing
