### 2026-01-11 #9542 @jordydavelaar
Hey! sorry for the delay, I will work on it this week

Jordy

> On Jan 11, 2026, at 8:06 AM, Arfon Smith ***@***.***> wrote:
>
>
> arfon
>  left a comment
> (openjournals/joss-reviews#9542)
>  <https://github.com/openjournals/joss-reviews/issues/9542#issuecomment-3734534697>
> 👋 @jordydavelaar <https://github.com/jordydavelaar> – happy new year! When do you think you might get started on your review?
>
