### 2026-02-04 b186ae5
add wishlist, autoversion, cleanup example scripts
