+++ examples/black_hole_accretion/black_hole_accretion.py
@@ -1,9 +1,6 @@
 import jax.numpy as jnp
 import argparse
-import sys
 import matplotlib.pyplot as plt
-
-sys.path.append("../../")  # XXX
 import jaxion
 
 # switch on for double precision
