+++ examples/logo_inverse_problem/logo_inverse_problem.py
@@ -1,8 +1,6 @@
 import jax
 import jax.numpy as jnp
-import sys
 
-sys.path.append("../../")  # XXX
 import jaxion
 import chex
 from typing import NamedTuple
