+++ examples/black_hole_accretion/plot.py
@@ -1,6 +1,6 @@
-import jaxion
 import jax.numpy as jnp
 import matplotlib.pyplot as plt
+import jaxion
 
 
 def main():
