+++ examples/soliton_gas_star/soliton_gas_star.py
@@ -1,8 +1,5 @@
 import jax.numpy as jnp
 import argparse
-import sys
-
-sys.path.append("../../")  # XXX
 import jaxion
 
 # switch on for double precision
