+++ examples/kinetic_condensation/kinetic_condensation.py
@@ -1,8 +1,8 @@
 import jax.numpy as jnp
 import numpy as np
 import jaxdecomp as jd
-import jaxion
 import argparse
+import jaxion
 
 # switch on for double precision
 # jax.config.update("jax_enable_x64", True)
