### 2026-04-05 #9542 @arfon
@chaitanyaafle @FrancescoFlammini @giovannifereoli - would any of you be willing to review this submission for JOSS? The submission under consideration is _Jaxion: A JAX package for Fuzzy Dark Matter_

Many thanks
Arfon
