### 2026-05-20 c729dc4
feat(domain): add rectangular grid support via aspect_ratio
Introduces an optional `aspect_ratio` domain parameter (default 1, fully
backward-compatible) so that nx = aspect_ratio * resolution while ny == nz
== resolution, keeping dx = dy = dz uniform. Enables economical wind-tunnel
simulations without any algorithmic changes to the integrator.

- defaults.json: new domain.aspect_ratio key (default 1)
- simulation.py: aspect_ratio / nx / shape properties; grid and kgrid build
  separate arrays per axis; state arrays use self.shape; particle volume
  corrected to Lx*Ly*Lz; num_cells and start-u
[…]
