+++ jaxion/simulation.py
@@ -58,6 +58,8 @@ class Simulation:
         # additional checks
         if self.resolution % 2 != 0:
             raise ValueError("Resolution must be divisible by 2.")
+        if self.nx % 2 != 0:
+            raise ValueError("nx (aspect_ratio * resolution) must be divisible by 2.")
 
         if self.params["time"]["adaptive"]:
             raise NotImplementedError("Adaptive time stepping is not yet implemented.")
@@ -113,22 +115,14 @@ class Simulation:
         if self.params["physics"]["cosmology"]:
             self.state["redshift"] = 0.0
         if self.params["physics"]["quantum"]:
-            self.state["psi"] = self.xzeros_jit(self.resolution) * 1j
+            self.state["psi"] = self.xzeros_jit(self.shape) * 1j
         if self.params["physics"]["external_potential"]:
-            self.state["V_ext"] = self.xzeros_jit(self.resolution)
+            self.state["V_ext"] = self.xzeros_jit(self.shape)
         if self.params["physics"]["hydro"]:
-            self.state["rho"] = jnp.zeros(
-                (self.resolution, self.resolution, self.resolution),
-            )
-            self.state["vx"] = jnp.zeros(
-                (self.resolution, self.resolution, self.resolution),
-            )
-            self.state["vy"] = jnp.zeros(
-                (self.resolution, self.resolution, self.resolution),
-            )
-            self.state["vz"] = jnp.zeros(
-                (self.resolution, self.resolution, self.resolution),
-            )
+            self.state["rho"] = jnp.zeros(self.shape)
+            self.state["vx"] = jnp.zeros(self.shape)
+            self.state["vy"] = jnp.zeros(self.shape)
+            self.state["vz"] = jnp.zeros(self.shape)
         if self.params["physics"]["particles"]:
             self.state["pos"] = jnp.zeros((self.num_particles, 3))
             self.state["vel"] = jnp.zeros((self.num_particles, 3))
@@ -154,12 +148,37 @@ class Simulation:
     @property
     def resolution(self):
         """
-        Return the (linear) resolution of the simulation
+        Return the (linear) resolution of the simulation (ny == nz)
         """
         return (
             self.params["domain"]["resolution_base"]
             * self.params["domain"]["resolution_multiplier"]
         )
+
+    @property
+    def aspect_ratio(self):
+        """
+        Return the aspect ratio of the simulation domain in the x direction.
+        nx = aspect_ratio * resolution; ny = nz = resolution.
+        Cell spacing dx = dy = dz is preserved.
+        """
+        return self.params["domain"]["aspect_ratio"]
+
+    @property
+    def nx(self):
+        """
+        Return the number of cells in the x direction (elongated axis).
+        """
+        return self.resolution * self.aspect_ratio
+
+    @property
+    def shape(self):
+        """
+        Return the grid shape (nx, ny, nz) as a tuple.
+        For a cubic domain this equals (resolution, resolution, resolution).
+        """
+        n = self.resolution
+        return (self.nx, n, n)
 
     @property
     def num_particles(self):
@@ -178,7 +197,8 @@ class Simulation:
     @property
     def dx(self):
         """
-        Return the cell size size of the simulation (kpc)
+        Return the (uniform) cell size of the simulation (kpc).
+        dx = dy = dz = box_size / resolution; Lx = aspect_ratio * box_size.
         """
         return self.box_size / self.resolution
 
[…]
