### 2026-05-26 JaxionProject/jaxion#3 @chaitanyaafle
[JOSS review] Installation fails on macOS arm64 (jaxdecomp builds from source, needs ninja/cmake)
Thanks for jaxion, it's a nice package and ran cleanly once I had it installed on a supported setup. Flagging this as part of the JOSS review (openjournals/joss-reviews#9542).
The documented `pip install jaxion` doesn't currently work on macOS arm64 (M3, recent macOS, Python 3.12 via conda). Two things happen:

1. A default conda env lands on Python 3.14, where the install can't resolve because tensorflow and jaxlib have no 3.14 wheels yet. (Separate point, noted in #4 .)
2. On Python 3.12 the resolve succeeds but the jaxdecomp build fails. `jaxdecomp==0.2.9` is published as an sdist only, so pip compiles it at install time, and the build errors out:

```
-- Looking for a CUDA compiler - NOTFOUND

fatal error: 'cstddef' file not found
ninja: build stopped: subcommand failed.
```

Worth being clear on what this is and isn't. CUDA being absent is fine, the build correctly falls back to the pure-JAX backend (`JD_JAX_BACKEND`), so this is not a GPU requirement. The real blockers are that `ninja` wasn't on the build path and the C++ headers aren't found on the current Command Line Tools and SDK (AppleClang 17). On Linux it builds and runs fine once `ninja` and `cmake` are present, verified on Colab.
Any of these would help:

* Note the supported platforms in the install docs, and that macOS may need to build jaxdecomp from source.
* List the build prerequisites (a C++ toolchain, `ninja`, `cmake`) for when a jaxdecomp wheel isn't available.
* Point macOS users to the Colab notebook or a Linux path as the easy option.

The install docs (https://jaxion.readthedocs.io/en/latest/pages/installation.html) list pip install jaxion and a build-from-source path, but don't mention supported platforms or that a C++ toolchain, ninja, and cmake are needed when no jaxdecomp wheel is available, which is the case on macOS arm64. Both documented commands compile jaxdecomp from source without flagging this. The missing macOS wheel is upstream in jaxdecomp, so the practical fix is probably to document the prerequisites and point macOS users to the Colab notebook or a Linux path.

EDIT: formatting + issue 4 link
