### 2026-05-27 47d1cd3
adding more installation details
