+++ docs/pages/quickstart.rst
@@ -7,7 +7,10 @@ To quickly try out Jaxion, install the package via pip:
 
     pip install jaxion
 
-And run a a simple example simulation (``examples/soliton_binary_merger/soliton_binary_merger.py``):
+If pip needs to build ``jaxdecomp`` from source, install the build prerequisites listed
+on the :doc:`Installation <installation>` page first.
+
+And run a simple example simulation (``examples/soliton_binary_merger/soliton_binary_merger.py``):
 
 .. literalinclude:: ../../examples/soliton_binary_merger/soliton_binary_merger.py
   :language: python
@@ -24,6 +27,8 @@ Running this should take under a minute and produce output (in ``checkpoints2/``
    :width: 480px
    :align: center
    :alt: soliton binary merger
+
+You can also run the example in the cloud: `Colab notebook <https://colab.research.google.com/github/JaxionProject/jaxion/blob/main/notebooks/soliton_binary_merger/soliton_binary_merger.ipynb>`__.
 
 
 For more info
