### 2026-05-27 JaxionProject/jaxion#3 @pmocz
Thank you for thoroughly trying out jaxion and tracking down these installation issues! I am adding some [documentation](https://github.com/JaxionProject/jaxion/commit/47d1cd311e6dcf6b3bb56e029aa1f1578a4396a6) to the installation docs
