### 2026-05-27 1cfc46e
update docs and reqs
