### 2026-05-27 ca647cc
particle support aspect ratio
