+++ pyproject.toml
@@ -5,7 +5,7 @@ authors = [
     { name = "Philip Mocz", email = "philip.mocz@gmail.com" },
 ]
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.10"
 license = "Apache-2.0"
 dynamic = ["version", "dependencies"]
 
